upgrade code

+++ src/main/java/com/jstn9/expensetracker/service/CategoryService.java
@@ -4,11 +4,13 @@ import com.jstn9.expensetracker.dto.category.CategoryCreateRequest;
 import com.jstn9.expensetracker.dto.category.CategoryResponse;
 import com.jstn9.expensetracker.dto.category.CategoryUpdateRequest;
 import com.jstn9.expensetracker.exception.CategoryAlreadyExistsException;
+import com.jstn9.expensetracker.exception.CategoryIsUsedInTransactionException;
 import com.jstn9.expensetracker.exception.CategoryNotFoundException;
 import com.jstn9.expensetracker.models.Category;
 import com.jstn9.expensetracker.models.User;
 import com.jstn9.expensetracker.repository.CategoryRepository;
-import com.jstn9.expensetracker.util.mapper.CategoryMapper;
+import com.jstn9.expensetracker.repository.TransactionRepository;
+import com.jstn9.expensetracker.util.MapperUtil;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -17,18 +19,20 @@ import java.util.List;
 public class CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final TransactionRepository transactionRepository;
     private final UserService userService;
 
-    public CategoryService(CategoryRepository categoryRepository, UserService userService) {
+    public CategoryService(CategoryRepository categoryRepository, TransactionRepository transactionRepository, UserService userService) {
         this.categoryRepository = categoryRepository;
+        this.transactionRepository = transactionRepository;
         this.userService = userService;
     }
 
     public List<CategoryResponse> getAllForCurrentUser(){
         User user = userService.getCurrentUser();
 
-        return categoryRepository.findByUser(user).stream()
-                .map(CategoryMapper::toCategoryResponse)
+        return categoryRepository.findByUserOrderByName(user).stream()
+                .map(MapperUtil::toCategoryResponse)
                 .toList();
     }
 
@@ -38,7 +42,7 @@ public class CategoryService {
         Category category = categoryRepository.findByIdAndUser(id, user)
                 .orElseThrow(() -> new CategoryNotFoundException("Category not found!"));
 
-        return CategoryMapper.toCategoryResponse(category);
+        return MapperUtil.toCategoryResponse(category);
     }
 
     public CategoryResponse createCategory(CategoryCreateRequest request){
@@ -53,7 +57,7 @@ public class CategoryService {
         category.setUser(user);
         Category savedCategory = categoryRepository.save(category);
 
-        return CategoryMapper.toCategoryResponse(savedCategory);
+        return MapperUtil.toCategoryResponse(savedCategory);
     }
 
     public CategoryResponse updateCategory(Long id, CategoryUpdateRequest request) {
@@ -65,7 +69,7 @@ public class CategoryService {
         category.setName(request.getName());
         Category savedCategory = categoryRepository.save(category);
 
-        return CategoryMapper.toCategoryResponse(savedCategory);
+        return MapperUtil.toCategoryResponse(savedCategory);
     }
 
     public void deleteCategory(Long id) {
@@ -73,6 +77,11 @@ public class CategoryService {
 
         Category category = categoryRepository.findByIdAndUser(id, user)
                 .orElseThrow(() -> new CategoryNotFoundException("Category not found!"));
+
+        boolean isUsed = transactionRepository.existsByCategory(category);
+        if(isUsed){
+            throw new CategoryIsUsedInTransactionException("Category is used in transactions and cannot be deleted!");
+        }
 
         categoryRepository.delete(category);
     }
